recorder/requirements.txt
psycopg2~=2.9.1
PyYAML~=6.0
pydantic~=1.8.2